% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/school.R
\name{school}
\alias{school}
\title{Get all of the data for a school}
\usage{
school(unitid)
}
\arguments{
\item{unitid}{The unitid for the particular institution}
}
\description{
This function gets the data for school, given its UNITID
}
\examples{
school(unitid=164924)
}

